test(device-quota): address phase 4 review findings
